split component-utils tests into their own dir
