+++ test/browser/component-utils/index.js
@@ -1,0 +1,2 @@
+import './controlled-app';
+import './proxy-component';
