+++ test/browser/component.js
@@ -481,69 +481,3 @@ describe(`Rendering exception`, function() {
     });
   });
 });
-
-describe(`Controlled App`, function() {
-  let el;
-
-  beforeEach(function(done) {
-    document.body.innerHTML = ``;
-    el = document.createElement(`controlled-app`);
-    document.body.appendChild(el);
-    window.requestAnimationFrame(() => done());
-  });
-
-  it(`does not allow update on component`, function() {
-    expect(() => el.update({foo: `not bar`})).to.throw(/update\(\) not allowed from component. Use controller/);
-  });
-
-  it(`Behaves like normal component`, function(done) {
-    let count = 0;
-    expect(el.controller.state).to.be.eql({count});
-    expect(el.textContent).to.contain(`Counter: ${count}`);
-
-    el.querySelector(`button.incr`).click();
-    count += 1;
-    expect(el.controller.state).to.be.eql({count});
-
-    window.requestAnimationFrame(() => {
-      expect(el.textContent).to.contain(`Counter: ${count}`);
-      done();
-    });
-  });
-});
-
-describe(`ProxyComponent`, function() {
-  let el;
-
-  beforeEach(function(done) {
-    document.body.innerHTML = ``;
-    el = document.createElement(`proxy-app`);
-    el.setAttribute(`data-foo`, `bar`);
-    document.body.appendChild(el);
-    window.requestAnimationFrame(() => done());
-  });
-
-  it(`renders the target element with attributes`, function() {
-    const child = document.querySelector(`event-producer`);
-    expect(child).to.not.eq(null);
-    expect(child.getAttribute(`data-foo`)).to.eq(`bar`);
-  });
-
-  it(`re-dispatches events from child that are not composed`, function(done) {
-    el.addEventListener(`nonComposedEvent`, () => done())
-    el.setAttribute(`send-non-composed`, ``);
-  });
-
-  it(`bubbles composed events`, function(done) {
-    el.addEventListener(`composedEvent`, () => done());
-    el.setAttribute(`send-composed`, ``);
-  });
-
-  it(`re-calculates the target on update`, function() {
-    el.setAttribute(`href`, `#local-link`);
-
-    expect(document.querySelector(`a`)).to.not.eq(null);
-    expect(document.querySelector(`event-producer`)).to.eq(null);
-  });
-});
-
